+++ client/src/common/fileManager/FileManagerPathBreadcrumbs.js
@@ -1,16 +1,55 @@
-import React from "react";
+/*
+ * Copyright 2019-2023 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React, {useMemo} from "react";
 import PropTypes from "prop-types";
 import {Breadcrumb} from "react-bootstrap";
-import {getPathParts} from "./fileManagerUtils";
 
 const FileManagerPathBreadcrumbs = ({
     rootFolder,
-    folder,
+    paths,
     handlePathUpdate
 }) => {
 
   const rootPath = rootFolder.path;
-  const paths = folder ? getPathParts(folder.path, rootPath) : [];
+
+  const breadcrumbs = useMemo(() => paths
+  .filter(path => !(path.index === 0 && path.path === "")) // empty first folder name
+  .filter(path => !!path.isBrowsable)
+  .map((path) => {
+    console.debug("Path", path);
+    if (path.isLast) {
+      return (
+          <Breadcrumb.Item key={path.index} active={true}>
+            {path.label}
+          </Breadcrumb.Item>
+      )
+    }
+    else {
+      return (
+          <Breadcrumb.Item
+              key={path.index}
+              onClick={() => handlePathUpdate(path.path)}
+              active={false}
+          >
+            {path.label}
+          </Breadcrumb.Item>
+      );
+    }
+  }), [paths, handlePathUpdate]);
 
   console.debug("Path breadcrumbs", paths);
 
@@ -21,37 +60,8 @@ const FileManagerPathBreadcrumbs = ({
           Home
         </Breadcrumb.Item>
 
-        {
-          paths
-          .filter(path => !(path.index === 0 && path.path === "")) // empty first folder name
-          .filter(path => !path.isRoot) // same as the root folder
-          .map((path) => {
-            if (path.isLast) {
-              return (
-                  <Breadcrumb.Item key={path.index} active={true}>
-                    {path.label}
-                  </Breadcrumb.Item>
-              )
-            }
-            else if (path.isChild) {
-              return (
-                  <Breadcrumb.Item
-                      key={path.index}
-                      onClick={() => handlePathUpdate(path.path)}
-                      active={false}
-                  >
-                    {path.label}
-                  </Breadcrumb.Item>
-              );
-            } else {
-              return (
-                  <Breadcrumb.Item key={path.index} active={false}>
-                    {path.label}
-                  </Breadcrumb.Item>
-              );
-            }
-          })
-        }
+        { breadcrumbs }
+
       </Breadcrumb>
   )
 
@@ -59,7 +69,7 @@ const FileManagerPathBreadcrumbs = ({
 
 FileManagerPathBreadcrumbs.propTypes = {
   rootFolder: PropTypes.object.isRequired,
-  folder: PropTypes.object,
+  paths: PropTypes.array.isRequired,
   handlePathUpdate: PropTypes.func.isRequired,
 }