+++ client/src/pages/assayForm/AssayForm.js
@@ -23,7 +23,8 @@ import {statuses} from "../../config/statusConstants";
 import UserInputs from "../../common/forms/UserInputs";
 import swal from 'sweetalert';
 import AssayTypeDropdown from "./AssayTypeDropdown";
-import CustomFieldCaptureInputList from "../../common/forms/customFields/CustomFieldCaptureInputList";
+import CustomFieldCaptureInputList
+  from "../../common/forms/customFields/CustomFieldCaptureInputList";
 import AttributeInputs from "../../common/forms/AttributeInputs";
 import {LoadingOverlay} from "../../common/loading";
 import ReactQuill from "react-quill";
@@ -34,9 +35,12 @@ import {Form as FormikForm, Formik} from "formik";
 import {useNavigate} from "react-router-dom";
 import * as yup from "yup";
 import axios from "axios";
-import FormikFormErrorNotification from "../../common/forms/FormikFormErrorNotification";
+import FormikFormErrorNotification
+  from "../../common/forms/FormikFormErrorNotification";
 import NotebookInputsCard from "../../common/forms/NotebookInputsCard";
-import TaskControlsDraggableCardList from "../../common/forms/tasks/TaskControlsDraggableCardList";
+import TaskControlsDraggableCardList
+  from "../../common/forms/tasks/TaskControlsDraggableCardList";
+import S3InputsCard from "../../common/forms/S3InputsCard";
 
 const AssayForm = props => {
 
@@ -505,6 +509,19 @@ const AssayForm = props => {
                       />
                     </Card.Body>
                   </Card>
+
+                  {/*S3*/}
+                  {
+                    !values.id ? (
+                        <S3InputsCard
+                            onChange={(key, value) => setFieldValue(key, value)}
+                            isActive={values.useS3}
+                            selectedStudy={study}
+                            errors={errors}
+                        />
+                    ) : ''
+
+                  }
 
                   {/* Attributes */}
                   <Card>