+++ web/src/main/java/io/studytracker/model/AssayOptions.java
@@ -27,6 +27,8 @@ public class AssayOptions {
   private String notebookTemplateId;
   private boolean useGit = false;
   private boolean useStorage = true;
+  private boolean useS3 = false;
+  private Long s3FolderId;
   private StorageDriveFolder parentFolder;
   private List<StorageDriveFolder> additionalFolders = new ArrayList<>();
   private GitGroup gitGroup;